feat: genMermaid
