+++ package/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./config.js";
 export * from "./result.js";
+export * from "./gen/mermaid.js";
